move rect rendering into rects.rs
